Feature: timetable 패키지내의 LearningTimeAggregateService에 소유자를 검증하는 로직 추가함 -#134

+++ _schedule/schedule/src/main/java/hwicode/schedule/timetable/application/LearningTimeAggregateService.java
@@ -1,8 +1,13 @@
 package hwicode.schedule.timetable.application;
 
+import hwicode.schedule.timetable.application.dto.learning_time.LearningTimeDeleteSubjectCommand;
+import hwicode.schedule.timetable.application.dto.learning_time.LearningTimeModifySubjectCommand;
+import hwicode.schedule.timetable.application.dto.learning_time.LearningTimeModifySubjectOfSubTaskCommand;
+import hwicode.schedule.timetable.application.dto.learning_time.LearningTimeModifySubjectOfTaskCommand;
 import hwicode.schedule.timetable.domain.LearningTime;
 import hwicode.schedule.timetable.domain.SubjectOfSubTask;
 import hwicode.schedule.timetable.domain.SubjectOfTask;
+import hwicode.schedule.timetable.exception.application.TimeTableForbiddenException;
 import hwicode.schedule.timetable.infra.limited_repository.LearningTimeFindRepository;
 import hwicode.schedule.timetable.infra.limited_repository.SubjectOfSubTaskFindRepository;
 import hwicode.schedule.timetable.infra.limited_repository.SubjectOfTaskFindRepository;
@@ -19,32 +24,44 @@ public class LearningTimeAggregateService {
     private final SubjectOfSubTaskFindRepository subjectOfSubTaskFindRepository;
 
     @Transactional
-    public boolean deleteSubject(Long learningTimeId) {
-        LearningTime learningTime = learningTimeFindRepository.findById(learningTimeId);
+    public boolean deleteSubject(LearningTimeDeleteSubjectCommand command) {
+        LearningTime learningTime = learningTimeFindRepository.findById(command.getLearningTimeId());
 
+        if (!learningTime.isOwner(command.getUserId())) {
+            throw new TimeTableForbiddenException();
+        }
         return learningTime.deleteSubject();
     }
 
     @Transactional
-    public String changeSubject(Long learningTimeId, String subject) {
-        LearningTime learningTime = learningTimeFindRepository.findById(learningTimeId);
+    public String changeSubject(LearningTimeModifySubjectCommand command) {
+        LearningTime learningTime = learningTimeFindRepository.findById(command.getLearningTimeId());
 
-        return learningTime.changeSubject(subject);
+        if (!learningTime.isOwner(command.getUserId())) {
+            throw new TimeTableForbiddenException();
+        }
+        return learningTime.changeSubject(command.getSubject());
     }
 
     @Transactional
-    public String changeSubjectOfTask(Long learningTimeId, Long subjectOfTaskId) {
-        LearningTime learningTime = learningTimeFindRepository.findById(learningTimeId);
-        SubjectOfTask subjectOfTask = subjectOfTaskFindRepository.findById(subjectOfTaskId);
+    public String changeSubjectOfTask(LearningTimeModifySubjectOfTaskCommand command) {
+        LearningTime learningTime = learningTimeFindRepository.findById(command.getLearningTimeId());
 
+        if (!learningTime.isOwner(command.getUserId())) {
+            throw new TimeTableForbiddenException();
+        }
+        SubjectOfTask subjectOfTask = subjectOfTaskFindRepository.findById(command.getSubjectOfTaskId());
         return learningTime.changeSubjectOfTask(subjectOfTask);
     }
 
     @Transactional
-    public String changeSubjectOfSubTask(Long learningTimeId, Long subjectOfSubTaskId) {
-        LearningTime learningTime = learningTimeFindRepository.findById(learningTimeId);
-        SubjectOfSubTask subjectOfSubTask = subjectOfSubTaskFindRepository.findById(subjectOfSubTaskId);
+    public String changeSubjectOfSubTask(LearningTimeModifySubjectOfSubTaskCommand command) {
+        LearningTime learningTime = learningTimeFindRepository.findById(command.getLearningTimeId());
 
+        if (!learningTime.isOwner(command.getUserId())) {
+            throw new TimeTableForbiddenException();
+        }
+        SubjectOfSubTask subjectOfSubTask = subjectOfSubTaskFindRepository.findById(command.getSubjectOfSubTaskId());
         return learningTime.changeSubjectOfSubTask(subjectOfSubTask);
     }
 }
